Removed TimeStamp from Event Watches
As we are only returning the most recent and a single event count the
time stamp as it was original designed does not make sense any more.
Furthermore it is not used in 4diac IDE. Furthermore, for accurate
tracing of what is happening the monitoring infrastructure was never
really helpfull and we have now other means. So removing improves
monitoring perfmance without loosing any features.